pyproject.toml
[project]
name = "newstome"
version = "0.0.1"
description = "Morning news digest to Telegram"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "google-genai>=0.8",
    "feedparser>=6.0",
    "trafilatura>=1.12",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rapidfuzz>=3.9",
    "python-telegram-bot>=21.6",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["newstome"]
